+++ src/components/List/listChangelogData.ts
@@ -15,6 +15,8 @@ export const changelogData: ChangelogData[] = [
     type: "Update",
     affects: ["Functionality"],
     notes: [
+      "Removes deprecated `button` value in `linkTypesArray`.",
+      "Replaces positional function arguments with object for `checkListChildrenError.`",
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
     ],
   },